Fix compiler warnings

+++ src/tremotesf/ui/screens/serversettings/serversettingsdialog.cpp
@@ -277,7 +277,7 @@ namespace tremotesf
         auto speedLimitsGroupBox = new QGroupBox(qApp->translate("tremotesf", "Limits"), this);
         auto speedLimitsGroupBoxLayout = new QGridLayout(speedLimitsGroupBox);
 
-        const int maxSpeedLimit = std::numeric_limits<uint>::max() / 1024;
+        const int maxSpeedLimit = static_cast<int>(std::numeric_limits<uint>::max() / 1024);
         //: In this context, 'k' prefix means SI prefix, i.e kB = 1000 bytes
         const QString suffix(qApp->translate("tremotesf", " kB/s"));
